Added styled components package, renamed calc.js to util.js
