+++ package.json
@@ -32,6 +32,7 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
     "react-native-paper": "^2.12.0",
     "react-navigation": "^3.3.2",
-    "react-navigation-slide-from-right-transition": "^1.0.4"
+    "react-navigation-slide-from-right-transition": "^1.0.4",
+    "styled-components": "^4.1.3"
   }
 }
