+++ src/api/FetchManager.js
@@ -1,7 +1,7 @@
 import { Alert } from 'react-native';
 
 import ClientFactory from './ClientFactory';
-import { getRandomElement } from './calc';
+import { getRandomElement } from './util';
 import { ARTCC_URL, STATUS_URL } from '../config/constants.json';
 
 export default class FetchManager {
